requirements.txt
google-cloud-aiplatform
google-genai
google-auth
pandas
openpyxl
pydantic
vertexai
